Now transmitting a message every 1ms
I measured the frequency of the yellow LED on the reciever as 500Hz, which is half the
transmission rate. (This makes sense because it toggles the LED when a message is recieved)
Data corruption seems minor - most messages get through untouched. To do a proper test, I'll
need to actually give each packet a number, and then measure the dropped packets and
corrupted packets, and give statistics on each.